--- a/include/protopuf/byte.h
+++ b/include/protopuf/byte.h
@@ -32,11 +32,6 @@ namespace pp {
     /// A byte (contiguous) sequence reference (no ownership).
     using bytes = std::span<std::byte>;
 
-    /// Returns the byte-distance between `begin(a)` and `begin(b)`.
-    inline constexpr std::size_t begin_diff(bytes a, bytes b) {
-        // `std::to_address` is used here for MSVC, ref to https://github.com/microsoft/STL/issues/1435
-        return static_cast<std::size_t>(std::to_address(a.begin()) - std::to_address(b.begin()));
-    }
 }
 
 #endif //PROTOPUF_BYTE_H