Hi,

I've noted terrible performance in debug mode. So, I started to dig what's the problem... and I found algorithmic no issues in your code. So, the performance degradation is probably related to templated code.

Any way, I did minor changed, which led to some performance boost.

Here is a benchmarks (in Release mode):

before:
```
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
BM_protopuf_encode            17.2 ns         17.2 ns     40928792
BM_protopuf_safe_encode       29.7 ns         29.7 ns     22518719
BM_protobuf_encode            68.0 ns         68.0 ns     10340438
BM_protopuf_decode             188 ns          188 ns      3728154
BM_protopuf_safe_decode        217 ns          217 ns      3232510
BM_protobuf_decode             244 ns          244 ns      2906093
```

After:
```
------------------------------------------------------------------
Benchmark                        Time             CPU   Iterations
------------------------------------------------------------------
BM_protopuf_encode            17.3 ns         17.3 ns     40937909
BM_protopuf_safe_encode       30.5 ns         30.5 ns     22250119
BM_protobuf_encode            69.3 ns         69.3 ns     10183935
BM_protopuf_decode             145 ns          145 ns      4840484
BM_protopuf_safe_decode        162 ns          162 ns      4311272
BM_protobuf_decode             243 ns          243 ns      2895603
```