@basiliscos 2025-07-08
tests fail due to `begin_diff` removal. Should I remove it from tests too?