@basiliscos 2025-07-09
> In CMake, `Debug` mode usually means `-O0 -g`. If you want debug information but also want some performance, you can use `RelWithDebInfo` mode instead of `Release` which usually means `-O2 -g` and enables lots of compiler optimizations (also you can try `-O1 -g` if you want more precise debug information).

That's true, however I want **my code** to be compiled in full debug mode without optimizations, and as the library is header-only it cannot be easily achieved.